Fixed an error with floating point in the pixel shader.

+++ DirectX11 Engine/DirectX11 Engine/Src/FX/MultipleTextureVertexShader.hlsl
@@ -16,14 +16,14 @@ struct VertexOut
 	float4 pos		: SV_POSITION;
 	float2 tex		: TEXCOORD0;
 	float3 normal	: NORMAL;
-	float textureIndex: TEXINDEX;
+	int textureIndex: TEXINDEX;
 };
 
 
 VertexOut VS(VertexIn vIn)
 {
 	VertexOut vOut;
-	vOut.textureIndex = vIn.pos.w;
+	vOut.textureIndex = (int)vIn.pos.w;
 	vIn.pos.w = 1.0f;
 	float4 pos = vIn.pos;
